+++ src/constants.php
@@ -3,7 +3,18 @@
 namespace Rubix\ML
 {
     /**
+     * The current version of the library.
+     *
+     * @internal
+     *
+     * @var string
+     */
+    const VERSION = '0.4.0';
+
+    /**
      * A small number used in substitution of 0.
+     *
+     * @internal
      *
      * @var float
      */
@@ -12,6 +23,8 @@ namespace Rubix\ML
     /**
      * The natural logarithm of the epsilon constant.
      *
+     * @internal
+     *
      * @var float
      */
     const LOG_EPSILON = -18.420680744;
@@ -19,17 +32,25 @@ namespace Rubix\ML
     /**
      * The number of radians in a full circle.
      *
+     * @internal
+     *
      * @var float
      */
     const TWO_PI = 2.0 * M_PI;
 
     /**
      * Half of pi.
+     *
+     * @internal
+     *
+     * @var float
      */
     const HALF_PI = 0.5 * M_PI;
 
     /**
      * Coefficient that determines floating point precision of random number generation.
+     *
+     * @internal
      *
      * @var int
      */